requirements.txt
azure-identity>=1.15.0
azure-core>=1.29.0
